// src/views/Update/types/UpdateTypes.ts
export enum DataFormat {
  RDFXML = "application/rdf+xml",
  NTRIPLES = "text/plain",
  TURTLE = "text/turtle",
  N3 = "text/rdf+n3",
  NQUADS = "text/x-nquads",
  JSONLD = "application/ld+json",
  RDFJSON = "application/rdf+json",
  TRIX = "application/trix",
  TRIG = "application/x-trig",
  RDF4JBINARY = "application/x-binary-rdf",
}